add arrow key function
