+++ frontend/src/views/DailyView/DailyView.scss
@@ -1,6 +1,3 @@
-body{
-    background-color: #48CAE4;
-}
 
 .daily-view-container{
 
@@ -50,12 +47,27 @@ margin-top: 0.1rem;
 color: #102a71;
 }
 
+.arrow-back:hover{
+    font-size: 3rem !important;
+    margin-right: 1rem !important;
+    margin-top: 0.1rem;
+    color: #FCD433 !important;
+    cursor: pointer;
+    }
+
 .arrow-forward{
     font-size: 3rem !important;
     margin-left: 1rem !important;
     color: #102a71;
     margin-top: 0.1rem;
 }
+.arrow-forward:hover{
+    font-size: 3rem !important;
+    margin-left: 1rem !important;
+    color: #FCD433;
+    margin-top: 0.1rem;
+    cursor: pointer;
+    }
 }
 
 @media (max-width: 830px) {
